% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{available_databases}
\alias{available_databases}
\title{List available databases supported by the package}
\usage{
available_databases()
}
\value{
Character vector of database names
}
\description{
List available databases supported by the package
}
