% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_and_create_dir}
\alias{check_and_create_dir}
\title{Check if output directory exists and create if not}
\usage{
check_and_create_dir(output_dir)
}
\arguments{
\item{output_dir}{Path to output directory}
}
\value{
Logical indicating success
}
\description{
Check if output directory exists and create if not
}
\keyword{internal}
