% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{.download_gds}
\alias{.download_gds}
\title{Download GDS (Dataset) data}
\usage{
.download_gds(
  accession,
  acc_dir,
  getGPL,
  force_download,
  max_retries,
  verbose,
  ...
)
}
\description{
Download GDS (Dataset) data
}
\keyword{internal}
