% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{.download_gpl}
\alias{.download_gpl}
\title{Download GPL (Platform) data}
\usage{
.download_gpl(
  accession,
  acc_dir,
  data_type,
  AnnotGPL,
  force_download,
  max_retries,
  verbose,
  ...
)
}
\description{
Download GPL (Platform) data
}
\keyword{internal}
