% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{.download_gse}
\alias{.download_gse}
\title{Download GSE (Series) data}
\usage{
.download_gse(
  accession,
  acc_dir,
  data_type,
  platform,
  samples,
  getGPL,
  AnnotGPL,
  extract_archives,
  filter_files,
  force_download,
  max_retries,
  parseCharacteristics,
  verbose,
  ...
)
}
\description{
Download GSE (Series) data
}
\keyword{internal}
