% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{.download_gsm}
\alias{.download_gsm}
\title{Download GSM (Sample) data}
\usage{
.download_gsm(
  accession,
  acc_dir,
  data_type,
  extract_archives,
  force_download,
  max_retries,
  verbose,
  ...
)
}
\description{
Download GSM (Sample) data
}
\keyword{internal}
