% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{.extract_expressionset_metadata}
\alias{.extract_expressionset_metadata}
\title{Extract metadata from single ExpressionSet}
\usage{
.extract_expressionset_metadata(
  eset,
  extract_pheno,
  extract_features,
  simplify
)
}
\description{
Extract metadata from single ExpressionSet
}
\keyword{internal}
