% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{.extract_gds_metadata}
\alias{.extract_gds_metadata}
\title{Extract metadata from GDS object}
\usage{
.extract_gds_metadata(gds_obj, extract_pheno)
}
\description{
Extract metadata from GDS object
}
\keyword{internal}
