% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{.extract_geo_archives}
\alias{.extract_geo_archives}
\title{Extract archives (tar.gz, gz, zip)}
\usage{
.extract_geo_archives(file_paths, verbose = TRUE)
}
\description{
Extract archives (tar.gz, gz, zip)
}
\keyword{internal}
