% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{.extract_gpl_metadata}
\alias{.extract_gpl_metadata}
\title{Extract metadata from GPL object}
\usage{
.extract_gpl_metadata(gpl_obj, extract_features)
}
\description{
Extract metadata from GPL object
}
\keyword{internal}
