% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{.extract_gse_list_metadata}
\alias{.extract_gse_list_metadata}
\title{Extract metadata from list of ExpressionSets}
\usage{
.extract_gse_list_metadata(geo_list, extract_pheno, extract_features, simplify)
}
\description{
Extract metadata from list of ExpressionSets
}
\keyword{internal}
