% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{.extract_gse_metadata}
\alias{.extract_gse_metadata}
\title{Extract comprehensive metadata from GSE}
\usage{
.extract_gse_metadata(gse_object, samples = NULL, verbose = TRUE)
}
\description{
Extract comprehensive metadata from GSE
}
\keyword{internal}
