% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{.extract_tcga_metadata}
\alias{.extract_tcga_metadata}
\title{Extract comprehensive metadata from TCGA SummarizedExperiment}
\usage{
.extract_tcga_metadata(se_object, verbose = TRUE)
}
\description{
Extract comprehensive metadata from TCGA SummarizedExperiment
}
\keyword{internal}
