% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{.identify_grouping_variables}
\alias{.identify_grouping_variables}
\title{Identify potential grouping variables in phenotype data}
\usage{
.identify_grouping_variables(pheno)
}
\description{
Identify potential grouping variables in phenotype data
}
\keyword{internal}
