% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{.save_tcga_data}
\alias{.save_tcga_data}
\title{Save TCGA data in various formats}
\usage{
.save_tcga_data(se_object, output_dir, project, format, verbose)
}
\description{
Save TCGA data in various formats
}
\keyword{internal}
