% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{.simplify_phenotype_data}
\alias{.simplify_phenotype_data}
\title{Simplify phenotype data by cleaning column names and values}
\usage{
.simplify_phenotype_data(pheno)
}
\description{
Simplify phenotype data by cleaning column names and values
}
\keyword{internal}
