% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtex.R
\name{download_gtex}
\alias{download_gtex}
\title{Download GTEx data}
\usage{
download_gtex(output_dir = "gtex_data", data_type = "counts")
}
\arguments{
\item{output_dir}{Directory to save downloaded data}

\item{data_type}{Type of data to download: "counts", "tpm", or "clinical"}
}
\value{
Path to downloaded files
}
\description{
Download GTEx data
}
