% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atlas.R
\name{download_hca}
\alias{download_hca}
\title{Download data from Human Cell Atlas}
\usage{
download_hca(project_id, output_dir = "hca_data", file_type = "h5ad")
}
\arguments{
\item{project_id}{HCA project ID}

\item{output_dir}{Directory to save downloaded data}

\item{file_type}{Type of file to download: "h5ad" or "loom"}
}
\value{
Path to downloaded files
}
\description{
Download data from Human Cell Atlas
}
