% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_atlas.R
\name{download_sc_atlas}
\alias{download_sc_atlas}
\title{Download data from Single Cell Expression Atlas}
\usage{
download_sc_atlas(accession, output_dir = "sc_data", file_type = "loom")
}
\arguments{
\item{accession}{Experiment accession number (e.g., "E-MTAB-5061")}

\item{output_dir}{Directory to save downloaded data}

\item{file_type}{Type of file to download: "loom", "rds", or "processed"}
}
\value{
Path to downloaded files
}
\description{
Download data from Single Cell Expression Atlas
}
