% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sra.R
\name{download_sra}
\alias{download_sra}
\title{Download data from SRA}
\usage{
download_sra(accession, output_dir = "sra_data", method = "fasterq")
}
\arguments{
\item{accession}{SRA accession number (e.g., "SRR1234567")}

\item{output_dir}{Directory to save downloaded data}

\item{method}{Download method: "fasterq" (default) or "prefetch"}
}
\value{
Path to downloaded files
}
\description{
Download data from SRA
}
