% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{extract_geo_metadata}
\alias{extract_geo_metadata}
\title{Extract metadata from GEO object}
\usage{
extract_geo_metadata(geo_object)
}
\arguments{
\item{geo_object}{GEO object from getGEO}
}
\value{
Data frame with metadata
}
\description{
Extract metadata from GEO object
}
