% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sra.R
\name{get_sra_metadata}
\alias{get_sra_metadata}
\title{Get SRA metadata}
\usage{
get_sra_metadata(accession)
}
\arguments{
\item{accession}{SRA accession number}
}
\value{
Data frame with metadata
}
\description{
Get SRA metadata
}
