% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{list_tcga_projects}
\alias{list_tcga_projects}
\title{List available TCGA projects}
\usage{
list_tcga_projects(include_count = TRUE, legacy = FALSE)
}
\arguments{
\item{include_count}{Include number of cases per project (default: TRUE)}

\item{legacy}{Use legacy archive (default: FALSE)}
}
\value{
Data frame with project information
}
\description{
Get a list of all available TCGA projects with descriptions.
}
\examples{
\dontrun{
# List all TCGA projects
projects <- list_tcga_projects()

# View cancer types
View(projects)
}
}
