% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{set_download_dir}
\alias{set_download_dir}
\title{Set default download directory}
\usage{
set_download_dir(path = "rnaseq_data")
}
\arguments{
\item{path}{Path to default download directory}
}
\description{
Set default download directory
}
