% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{download_geo}
\alias{download_geo}
\title{Download data from GEO}
\usage{
download_geo(accession, output_dir = "geo_data", data_type = "matrix", ...)
}
\arguments{
\item{accession}{GEO accession number (e.g., "GSE12345")}

\item{output_dir}{Directory to save downloaded data}

\item{data_type}{Type of data to download: "matrix" (expression matrix) or "raw" (raw data)}

\item{...}{Additional parameters passed to getGEO or getGEOSuppFiles}
}
\value{
Path to downloaded files
}
\description{
Download data from GEO
}
