% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga.R
\name{download_tcga}
\alias{download_tcga}
\title{Download data from TCGA}
\usage{
download_tcga(
  project,
  output_dir = "tcga_data",
  data_category = "Transcriptome Profiling",
  data_type = "Gene Expression Quantification",
  workflow_type = "HTSeq - Counts"
)
}
\arguments{
\item{project}{TCGA project ID (e.g., "TCGA-BRCA")}

\item{output_dir}{Directory to save downloaded data}

\item{data_category}{Data category to download (default: "Transcriptome Profiling")}

\item{data_type}{Data type to download (default: "Gene Expression Quantification")}

\item{workflow_type}{Workflow type (default: "HTSeq - Counts")}
}
\value{
Path to downloaded files
}
\description{
Download data from TCGA
}
